feat: jwt 1. api.thrift 补充了一些信息，加入了 user_info_upload 2. 移除了 user 模块不必要的 token 参数 3. 更正了 hertz 框架的模块 4. 补充了 pkg 里的一些常量 5. 增加了一些额外中间件 6. 增加了 jwt 中间件，配置了相关配置项

+++ cmd/api/biz/handler/handler.go
@@ -7,7 +7,7 @@ import (
 )
 
 type Response struct {
-	StatusCode int64  `json:"status_code"`
+	StatusCode int32  `json:"status_code"`
 	StatusMsg  string `json:"status_msg,omitempty"`
 }
